+++ lib/common/page/dialog_page.dart
